Use top-level cctools when available
cctools was updated and migrated to the `by-name` hierarchy in nixpkgs,
which moves it to the top-level. It is also being added to
`darwin-aliases.nix`, which will make the old name unavailable for use
in nixpkgs. This change preferentially uses the new name while falling
back to the old one for out-of-tree users.

Relevant nixpkgs PRs:
- https://github.com/NixOS/nixpkgs/pull/307880
- https://github.com/NixOS/nixpkgs/pull/328077